# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API - Carga Pesada",
        "description": "API RESTFul Carga Pesada",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/contacts": {
            "get": {
                "tags": [
                    "Contact"
                ],
                "summary": "Listar todos os contatos",
                "operationId": "852ed2762dbbe0a94ccfc4c97587621a",
                "responses": {
                    "200": {
                        "description": "Lista de contatos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Contact"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Contact"
                ],
                "summary": "Criar um novo contato",
                "operationId": "379869295076c3201047aea1a0ece726",
                "requestBody": {
                    "description": "Dados do novo contato",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nome": {
                                        "description": "Nome do contato",
                                        "type": "string"
                                    },
                                    "phone": {
                                        "description": "Número de telefone do contato",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Contato criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Contato criado com sucesso."
                                        },
                                        "contact": {
                                            "$ref": "#/components/schemas/Contact"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/contacts/{id}": {
            "get": {
                "tags": [
                    "Contact"
                ],
                "summary": "Obter detalhes de um contato específico",
                "operationId": "b50e403e9334c34bf4a1d319a863d3e7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do contato",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do contato",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Contact"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Contato não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Contato não encontrado."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Contact"
                ],
                "summary": "Atualizar um contato existente",
                "operationId": "a6027289e451f3cd7ddfec1b24203aae",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do contato a ser atualizado",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Dados do contato a ser atualizado",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreUpdateContact"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Contato atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Contact"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Contato não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Contact"
                ],
                "summary": "Excluir um contato",
                "operationId": "564a2f7f26488abf9de14d86fac81557",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do contato a ser excluído",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Contato excluído com sucesso"
                    },
                    "404": {
                        "description": "Contato não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Contact": {
                "required": [
                    "id",
                    "nome",
                    "phone"
                ],
                "properties": {
                    "id": {
                        "description": "ID único do contato (UUID)",
                        "type": "string",
                        "format": "uuid"
                    },
                    "nome": {
                        "description": "Nome do contato",
                        "type": "string"
                    },
                    "phone": {
                        "description": "Número de telefone do contato",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "StoreUpdateContact": {
                "title": "StoreUpdateContact",
                "required": [
                    "nome",
                    "phone"
                ],
                "properties": {
                    "nome": {
                        "description": "Nome do contato",
                        "type": "string"
                    },
                    "phone": {
                        "description": "Número de telefone do contato",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}